cleanup: split ea-eaac formats
